requirements.txt
Flask
openai
beautifulsoup4
requests
python-dotenv
flask-cors